Run tests on BrowserStack
 - Moved the original `karma.conf.js` to karma-saucelabs.conf.js
 - load the correct conf by inspecting the env.
 - pick default local browsers based on `os.platform`
 - TravisCI uses the SauceLabs config
 - device-list.json has a full run of popular devices
     - international
